add deleteItem function + edit my diffUtil

+++ app/src/main/java/com/example/shop/presentation/entity/ItemUi.kt
@@ -2,13 +2,13 @@ package com.example.shop.presentation.entity
 
 import com.example.shop.domain.entity.ItemDomain
 
-data class ItemUi (
-    val id:Int,
-    val title:String,
+data class ItemUi(
+    val id: Long,
+    val title: String,
     val marked: Boolean
 )
 
-fun ItemUi.toDomain(): ItemDomain{
+fun ItemUi.toDomain(): ItemDomain {
     return ItemDomain(
         id = this.id,
         title = this.title,
@@ -16,7 +16,7 @@ fun ItemUi.toDomain(): ItemDomain{
     )
 }
 
-fun ItemDomain.toUi(): ItemUi{
+fun ItemDomain.toUi(): ItemUi {
     return ItemUi(
         id = this.id,
         title = this.title,
